feat: implement Haml and RuboCop linting functions and integrate into formatter
