+++ lib/list_cops.rb
@@ -1,21 +1,28 @@
-require 'rubocop'
-require 'haml_lint'
 require 'json'
 
 def rubocop_cops
+  require 'rubocop'
+
   config_store = RuboCop::ConfigStore.new
   config = config_store.for('.').to_h
 
   cop_list = [
     'Style/StringLiterals',
+    'Layout/SpaceInsideParens',
   ]
 
   config.select { |cop, _| cop_list.include?(cop) }
+rescue StandardError
+  {}
 end
 
 def haml_lint_cops
+  require 'haml_lint'
+
   config = HamlLint::ConfigurationLoader.load_applicable_config
   config.hash
+rescue StandardError
+  {}
 end
 
 def list_cops
